refactor: move api to livingdocs_api
